refactor(rapid): split change propagation along module boundaries
log_populate.cpp now owns the propagation machinery and nothing about any
particular record format: the coordinator and per-table workers, the ring
buffers and apply watermarks, and the retry/quarantine policy. The
per-source switch in the worker collapses to a routing call.

Each source now owns its own record contract, exposed as
apply_change()/validate_record(). RedoLog::LogParser lives in
log_redolog.{h,cc} and DML::CopyInfoParser in log_dml_notification.{h,cc}
(renamed from log_copyinfo.{h,cc}, which had grown past its name).
apply_change() prepares Rapid_load_context, parses and applies the record,
classifies the outcome and, for COPY_INFO, resolves the primary
COMMIT/ROLLBACK race before reporting the record to TransactionManager. It
returns the new ChangeApplyResult (log_commons.h): APPLIED / RETRYABLE /
PERMANENT plus parsed_bytes and the stale_reason_t to record if the table is
quarantined, so the worker no longer re-derives failure semantics from
Source.

The capture-side helpers (RegisterCopyInfoParticipant, EnqueueCopyInfo)
move out of ha_shannon_rapid.cc into log_dml_notification and up into
Populate, so the redo path can reuse them.

TransactionManager -- COPY_INFO transaction participation and the
asynchronous outcome rendezvous -- moves with the format it serves, out of
log_populate into log_dml_notification. PopulatorImpl still drives its
start/forget/shutdown alongside the propagation thread lifecycle, which is
where that orchestration belongs.

Also fold LoadedTables::add/get/erase back into the header where they are
declared, return early from read_off_page_data() when the table has no blob
fields, drop the unused Populator::load_indexes_caches() declaration and
call site, and mark the partitioning columns readable with ColumnMapGuard in
resolve_change_partitions() instead of touching the read_set by hand.
