+++ storage/rapid_engine/populate/log_dml_notification.h
@@ -27,17 +27,111 @@
    The fundmental code for imcs. The chunk is used to store the data which
    transfer from row-based format to column-based format.
 */
-#ifndef __SHANNONBASE_LOG_COPY_INFO_PARSER_H__
-#define __SHANNONBASE_LOG_COPY_INFO_PARSER_H__
+#ifndef __SHANNONBASE_LOG_DML_NOTIFICATION_H__
+#define __SHANNONBASE_LOG_DML_NOTIFICATION_H__
+
+#include <atomic>
+#include <cstdint>
+#include <mutex>
+#include <unordered_map>
+#include <unordered_set>
+#include <vector>
 
 #include "sql/sql_base.h"
 #include "storage/rapid_engine/include/rapid_const.h"
 #include "storage/rapid_engine/include/rapid_context.h"  //Rapid_load_context
 #include "storage/rapid_engine/populate/log_commons.h"   //change_record_buff_t::OperType
+#include "storage/rapid_engine/trx/transaction.h"        //TransactionSubscriber
 #include "storage/rapid_engine/utils/utils.h"
 
 namespace ShannonBase {
 namespace Populate {
+
+/**
+ * Owns COPY_INFO transaction participation and the asynchronous transaction
+ * outcome rendezvous used by the propagation workers.
+ *
+ * Synchronous state is keyed by THD while a source transaction is active.
+ * Asynchronous state is keyed by the captured InnoDB writer transaction id and
+ * may outlive the THD-local Transaction facade until every queued record is
+ * applied.
+ */
+class TransactionManager final : public TransactionSubscriber {
+ public:
+  enum class Outcome : uint8_t { ACTIVE = 0, COMMITTED, ABORTED };
+
+  struct Registration {
+    Transaction::ID source_trx_id{0};
+
+    explicit operator bool() const noexcept { return source_trx_id != 0; }
+  };
+
+  static TransactionManager &instance() {
+    static TransactionManager manager;
+    return manager;
+  }
+
+  TransactionManager(const TransactionManager &) = delete;
+  TransactionManager &operator=(const TransactionManager &) = delete;
+
+  Registration register_change(THD *thd, table_id_t table_id);
+
+  // ROLLBACK TO SAVEPOINT is a partial transaction undo, not a normal
+  // statement-rollback callback. Keep this explicit while COPY_INFO lacks
+  // per-operation undo.
+  void quarantine_partial_rollback(THD *thd, const char *reason);
+
+  Outcome get_outcome(Transaction::ID txn_id, uint64_t *commit_scn = nullptr);
+  void on_change_applied(Transaction::ID txn_id, table_id_t table_id);
+  void forget_table(table_id_t table_id);
+
+  void on_transaction_commit(THD *thd) override;
+  void on_transaction_rollback(THD *thd) override;
+  void on_statement_commit(THD *thd) override;
+  void on_statement_rollback(THD *thd) override;
+  void on_transaction_detach(THD *thd) override;
+
+  void start();
+  void shutdown();
+
+ private:
+  struct Participant {
+    Transaction::ID source_trx_id{0};
+    std::unordered_set<table_id_t> touched_tables;
+    bool statement_has_changes{false};
+    bool fail_closed{false};
+  };
+
+  struct TableProgress {
+    uint64_t registered{0};
+    uint64_t applied{0};
+  };
+
+  struct TxnProgress {
+    Outcome outcome{Outcome::ACTIVE};
+    uint64_t commit_scn{0};
+    std::unordered_map<table_id_t, TableProgress> tables;
+  };
+
+  TransactionManager() = default;
+
+  void ensure_subscribed();
+  void clear();
+  void publish_commit(Transaction::ID txn_id, uint64_t commit_scn);
+  void publish_rollback(Transaction::ID txn_id);
+  void quarantine_participant(THD *thd, bool require_statement_change, const char *reason);
+  static void finalize_table(Transaction::ID txn_id, table_id_t table_id, Outcome outcome, uint64_t commit_scn);
+  void erase_if_complete_locked(Transaction::ID txn_id);
+
+  std::mutex m_subscription_mutex;
+  std::atomic<bool> m_subscribed{false};
+
+  std::mutex m_mutex;
+  std::unordered_map<THD *, Participant> m_participants;
+  std::unordered_map<Transaction::ID, TxnProgress> m_transactions;
+};
+
+namespace DML {
 /**
  * To parse the copy_info, it used to populate the changes from ionnodb
  * to rapid.
@@ -124,6 +218,25 @@ class CopyInfoParser {
   uint parse_copy_info(Rapid_load_context *context, table_id_t &table_id, change_record_buff_t::OperType oper_type,
                        byte *start, byte *end_ptr, byte *new_start, byte *new_end_ptr);
 
+  /**
+   * Apply one buffered COPY_INFO change record.
+   *
+   * Owns everything specific to the row-image source: validating the primary
+   * writer identity, resolving the asynchronous primary COMMIT/ROLLBACK race,
+   * preparing the Rapid_load_context and driving parse_copy_info(). On success
+   * it also reports the record to TransactionManager so the source transaction
+   * can be finalized. The propagation worker only routes the record here and
+   * acts on the returned status.
+   */
+  ChangeApplyResult apply_change(Rapid_load_context &context, change_record_buff_t &record);
+
+  /**
+   * Producer-side check: does this buffered record satisfy the COPY_INFO
+   * format invariants? Row-image propagation must carry the real primary
+   * InnoDB creator id (commit_scn == 0 is a valid ACTIVE version).
+   */
+  static bool validate_record(const change_record_buff_t &record);
+
  private:
   /**
    * @brief Apply an UPDATE operation to a RAPID table using COPY_INFO data.
@@ -202,6 +315,63 @@ class CopyInfoParser {
    */
   int parse_and_apply_delete(Rapid_load_context *context, table_id_t &table_id, const byte *start, const byte *end_ptr);
 };
+
+/**
+ * @brief hton se_after_commit: publish the Rapid-side transaction.
+ * @param[in] arg  the server's Trans_param.
+ */
+void rapid_after_commit(void *arg);
+
+/**
+ * @brief hton se_before_rollback: undo the Rapid-side transaction.
+ * @param[in] arg  the server's Trans_param.
+ */
+void rapid_before_rollback(void *arg);
+}  // namespace DML
+
+/**
+ * @brief Capture side of the change-propagation path.
+ *
+ * A change source builds a change record out of the row images it holds and
+ * hands it over to EnqueueCopyInfo() below: today the primary engine's DML
+ * notifier (ha_shannon_rapid.cc: NotifyAfterInsert / NotifyAfterUpdate /
+ * NotifyAfterDelete), and in the future the redo-log parser as well. Everything
+ * after that point - the transaction identity stamped on the record, the queue
+ * it enters and the primary COMMIT/ROLLBACK outcome that publishes or undoes it
+ * - is owned here, so a source is left with nothing but building row images.
+ *
+ * These two deliberately live in Populate rather than Populate::DML so that the
+ * redo-log parsing path can reuse them without depending on the COPY_INFO
+ * specific machinery.
+ */
+
+/**
+ * @brief Register Rapid as a statement participant of the current THD, and for
+ *        an explicit transaction also as a final transaction participant.
+ *
+ * Must run before the first change record of the statement is enqueued: the
+ * registration is what makes the server call the se_after_commit /
+ * se_before_rollback hooks (rapid_after_commit / rapid_before_rollback) once the
+ * primary transaction ends, which is what publishes the captured changes.
+ *
+ * @param[in] thd  thread whose transaction takes part in the propagation.
+ */
+void RegisterCopyInfoParticipant(THD *thd);
+
+/**
+ * @brief Stamp the transaction identity on a change record and enqueue it for
+ *        the population worker.
+ *
+ * The record leaves this call carrying the real primary InnoDB creator id and
+ * commit_scn == 0, i.e. it enters Rapid MVCC as an ACTIVE notification version;
+ * the outcome published on primary COMMIT/ROLLBACK is what finalizes it.
+ *
+ * @param[in]     thd     thread the DML statement runs on.
+ * @param[in,out] record  change record to enqueue; consumed on success.
+ * @return true if the record was enqueued, false if the thread has no usable
+ *         transaction and the change could not be captured.
+ */
+bool EnqueueCopyInfo(THD *thd, change_record_buff_t &&record);
 }  // namespace Populate
 }  // namespace ShannonBase
-#endif  //__SHANNONBASE_LOG_COPY_INFO_PARSER_H__
+#endif  //__SHANNONBASE_LOG_DML_NOTIFICATION_H__
